Added Parser for getting all results in one place

+++ UniversalSolution/FormulasCollection/Realizations/ConverterFormulas.cs
@@ -5,20 +5,20 @@ namespace FormulasCollection.Realizations
 {
     public class ConverterFormulas : IConverterFormulas
     {
-        public decimal IncorrectAmericanOdds => -127;
+        public double IncorrectAmericanOdds => -127;
 
-        public decimal ConvertAmericanToDecimal(int? american) => american == null && !american.HasValue && american.Value != 0
+        public double ConvertAmericanToDecimal(int? american) => american == null && !american.HasValue && american.Value != 0
                 ? IncorrectAmericanOdds
                 : (american.Value > 0
                     ? PositiveConvertationFormula(american.Value)
                     : NegativeConvertationFormula(american.Value));
 
 
-        public decimal PositiveConvertationFormula(int american) => american <= 0
+        public double PositiveConvertationFormula(int american) => american <= 0
             ? IncorrectAmericanOdds
             : (american / 100) + 1;
 
-        public decimal NegativeConvertationFormula(int american) => american >= 0
+        public double NegativeConvertationFormula(int american) => american >= 0
             ? IncorrectAmericanOdds
             : (100 / Math.Abs(american)) + 1;
     }
